service/postgres-to-kafka/requirements.txt
confluent-kafka==2.3.0
psycopg2==2.9.9
python-dotenv==1.0.1
